Update introduction.txt
At two separate points in my life I've been bitten by assuming the library was thread-safe under apparently read-only operations, at the cost of several hour's work. Both times, I couldn't find clear examples of the things I wanted to do (such as bulk-loading), and assumed the issue was a mistake on my end. In the end the solution was easy: use a mutex.

Looking through the git issues, it seems I'm not the only one who's had this experience. Putting a little notice here (and maybe elsewhere as well) seems like a low-cost way of heading off this kind of problem.